distribution/all_requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
autopep8==2.0.0
Babel==2.12.1
billiard==3.6.4.0
cachetools==5.2.1
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
daphne==4.0.0
Django==4.1.3
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-debug-toolbar==4.0.0
django-elasticsearch-dsl==7.2.2
django-elasticsearch-dsl-drf==0.22.5
django-environ==0.9.0
django-filter==22.1
django-money==3.0.0
django-nine==0.2.5
django-redis==5.2.0
django-sse==0.4.1
django-toggle-switch-widget==0.1.1
djangorestframework==3.14.0
drf-yasg==1.21.5
elasticsearch==7.17.7
elasticsearch-dsl==7.4.0
et-xmlfile==1.1.0
facebook-sdk==3.1.0
google-api-core==2.11.0
google-api-python-client==2.72.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.58.0
httplib2==0.21.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
Markdown==3.4.1
MarkupSafe==2.1.2
msgpack==1.0.5
mysqlclient==2.1.1
openpyxl==3.1.2
packaging==23.0
phonenumbers==8.13.4
pika==1.3.1
Pillow==9.3.0
polars==0.17.8
prompt-toolkit==3.0.36
protobuf==4.21.12
psycopg2==2.9.5
py-moneyed==2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
redis==4.5.5
requests==2.28.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
sse==1.2
tomli==2.0.1
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.3.0
xlsx2csv==0.8.1
XlsxWriter==3.0.9
zope.interface==6.0
